all: Open source SX1280 time-on-air calculation
